SIFTS

SIFTS

MIToS.SIFTS — Module.

The SIFTS module of MIToS allows to obtain the residue-level mapping between databases stored in the SIFTS XML files. It makes easy to assign PDB residues to UniProt/Pfam positions. Given the fact that pairwise alignments can lead to misleading association between residues in both sequences, SIFTS offers more reliable association between sequence and structure residue numbers.

Features

Download and parse SIFTS XML files

Store residue-level mapping in Julia

Easy generation of Dicts between residues numbers

using MIToS.SIFTS

source

Contents

SIFTS

Contents

Types

Constants

Macros

Methods and functions

Types

MIToS.SIFTS.SIFTSResidue — Type.

A SIFTSResidue object stores the SIFTS residue level mapping for a residue. It has the following fields that you can access at any moment for query purposes:

- `PDBe` : A `dbPDBe` object, it's present in all the `SIFTSResidue`s.
- `UniProt` : A `Nullable` wrapping a `dbUniProt` object.
- `Pfam` : A `Nullable` wrapping a `dbPfam` object.
- `NCBI` : A `Nullable` wrapping a `dbNCBI` object.
- `InterPro` : A `Nullable` wrapping a `dbInterPro` object.
- `PDB` : A `Nullable` wrapping a `dbPDB` object.
- `SCOP` : A `Nullable` wrapping a `dbSCOP` object.
- `CATH` : A `Nullable` wrapping a `dbCATH` object.
- `missing` : It's `true` if the residue is missing, i.e. not observed, in the structure.
- `sscode` : A string with the secondary structure code of the residue.
- `ssname` : A string with the secondary structure name of the residue.

source

MIToS.SIFTS.dbCATH — Type.

dbCATH stores the residue id, number, name and chain in CATH as strings.

source

MIToS.SIFTS.dbInterPro — Type.

dbInterPro stores the residue id, number, name and evidence in InterPro as strings.

source

MIToS.SIFTS.dbNCBI — Type.

dbNCBI stores the residue id, number and name in NCBI as strings.

source

MIToS.SIFTS.dbPDB — Type.

dbPDB stores the residue id, number, name and chain in PDB as strings.

source

MIToS.SIFTS.dbPDBe — Type.

dbPDBe stores the residue number and name in PDBe as strings.

source

MIToS.SIFTS.dbPfam — Type.

dbPfam stores the residue id, number and name in Pfam as strings.

source

MIToS.SIFTS.dbSCOP — Type.

dbSCOP stores the residue id, number, name and chain in SCOP as strings.

source

MIToS.SIFTS.dbUniProt — Type.

dbUniProt stores the residue id, number and name in UniProt as strings.

source

Constants

Macros

Methods and functions

Base.parse — Method.

parse(document::LightXML.XMLDocument, ::Type{SIFTSXML}; chain=All, missings::Bool=true)

Returns a Vector{SIFTSResidue} parsed from a SIFTSXML file. By default, parses all the chains and includes missings residues.

source

MIToS.SIFTS.downloadsifts — Method.

Download the gzipped SIFTS xml for the pdbcode. The extension of the downloaded file is .xml.gz by default. The filename can be changed, but the .xml.gz at the end is mandatory.

source

MIToS.SIFTS.siftsmapping — Method.

Parses a SIFTS XML file and returns a Dict between residue numbers of two DataBases with the given identifiers. A chain could be specified (All by default). If missings is true (default) all the residues are used, even if they haven’t coordinates in the PDB file.

source