Utils

MIToS.Utils — Module

The Utils has common utils functions and types used in other modules.

using MIToS.Utils

Contents

Utils

Contents

Types

Constants

Macros

Methods and functions

Types

MIToS.Utils.All — Type

All is used instead of MIToS 1.0 "all" or "*", because it's possible to dispatch on it.

MIToS.Utils.FileFormat — Type

FileFormat is used for write special parse (and read) methods on it.

Constants

MIToS.Utils.THREE2ONE — Constant

THREE2ONE is a dictionary that maps three-letter amino acid residue codes (String) to their corresponding one-letter codes (Char). The dictionary is generated by parsing components.cif file from the Protein Data Bank.

julia> using MIToS.Utils

julia> one_letter_code = THREE2ONE["ALA"]
'A': ASCII/Unicode U+0041 (category Lu: Letter, uppercase)

Macros

Methods and functions

Base.read — Method

read(pathname, FileFormat [, Type [, … ] ] ) -> Type

This function opens a file in the pathname and calls parse(io, ...) for the given FileFormat and Type on it. If the pathname is an HTTP or FTP URL, the file is downloaded with download in a temporal file. Gzipped files should end on .gz.

Base.write — Method

write{T<:FileFormat}(filename::AbstractString, object, format::Type{T}, mode::ASCIIString="w")

This function opens a file with filename and mode (default: "w") and writes (print) the object with the given format. Gzipped files should end on .gz.

MIToS.Utils.check_file — Method

Returns the filename. Throws an ErrorException if the file doesn't exist, or a warning if the file is empty.

MIToS.Utils.check_pdbcode — Method

It checks if a PDB code has the correct format.

MIToS.Utils.download_file — Method

download_file uses HTTP.jl to download files from the web. It takes the file url as first argument and, optionally, a path to save it. Keyword arguments are are directly passed to to HTTP.download (HTTP.request). Use the headers keyword argument to pass a Dict{String,String} with the header information. Set the HTTPS_PROXY and HTTPS_PROXY ENViromental variables if you are behind a proxy.

julia> using MIToS.Utils

julia> download_file("http://www.uniprot.org/uniprot/P69905.fasta","seq.fasta",
       headers = Dict("User-Agent" =>
                      "Mozilla/5.0 (compatible; MSIE 7.01; Windows NT 5.0)"))
"seq.fasta"

MIToS.Utils.get_n_words — Method

get_n_words{T <: Union{ASCIIString, UTF8String}}(line::T, n::Int) It returns a Vector{T} with the first n (possibles) words/fields (delimited by space or tab). If there is more than n words, the last word returned contains the finals words and the delimiters. The length of the returned vector is n or less (if the number of words is less than n). This is used for parsing the Stockholm format.

julia> using MIToS.Utils

julia> get_n_words("#=GR O31698/18-71 SS    CCCHHHHHHHHHHHHHHHEEEEEEEEEEEEEEEEHHH", 3)
3-element Vector{String}:
 "#=GR"
 "O31698/18-71"
 "SS    CCCHHHHHHHHHHHHHHHEEEEEEEEEEEEEEEEHHH"

MIToS.Utils.getarray — Method

Getter for the array field of NamedArrays

MIToS.Utils.hascoordinates — Method

hascoordinates(id) It returns true if id/sequence name has the format: UniProt/start-end (i.e. O83071/192-246)

MIToS.Utils.isnotemptyfile — Method

Returns true if the file exists and isn't empty.

MIToS.Utils.lineiterator — Method

Create an iterable object that will yield each line from a stream or string.

MIToS.Utils.list2matrix — Method

Returns a square symmetric matrix from the vector vec. side is the number of rows/columns. The diagonal is not included by default, set to true if there are diagonal elements in the list.

MIToS.Utils.matrix2list — Method

Returns a vector with the part ("upper" or "lower") of the square matrix mat. The diagonal is not included by default.

MIToS.Utils.select_element — Method

Selects the first element of the vector. This is useful for unpacking one element vectors. Throws a warning if there are more elements. element_name is element by default, but the name can be changed using the second argument.